+++ skbuild/command/egg_info.py
@@ -14,9 +14,11 @@ class egg_info(set_build_base_mixin, new_style(_egg_info)):
     """Custom implementation of ``egg_info`` setuptools command."""
 
     def finalize_options(self):
+        # pylint:disable=access-member-before-definition
         if self.egg_base is not None:
             script_path = os.path.abspath(self.distribution.script_name)
             script_dir = os.path.dirname(script_path)
+            # pylint:disable=attribute-defined-outside-init
             self.egg_base = os.path.join(script_dir, self.egg_base)
 
         super(egg_info, self).finalize_options()