+++ CHANGES.rst
@@ -8,6 +8,21 @@ details, see the commit logs at http://github.com/scikit-build/scikit-build
 Next Release
 ============
 
+Documentation
+-------------
+
+* Improve internal API documentation:
+
+  * :mod:`skbuild.platform_specifics.windows`
+  * :mod:`skbuild.command`
+  * :mod:`skbuild.command.generate_source_manifest`
+  * :mod:`skbuild.utils`
+
+Cleanups
+--------
+
+* Fix miscellaneous pylint warnings.
+
 Scikit-build 0.6.1
 ==================